Change internal substep representation to uint8_t
This decreases the serialized size of the class, although not the size
in memory because of padding and alignment.

Changing the exposed type to uint8_t would be a bad idea, because
uint8_t is a character type and so doesn't stream properly.

+++ src/Time/TimeStepId.hpp
@@ -64,7 +64,7 @@ class TimeStepId {
 
   int64_t slab_number_{std::numeric_limits<int64_t>::lowest()};
   Time step_time_{};
-  uint64_t substep_{0};
+  uint8_t substep_{0};
   Rational step_size_{};
   double substep_time_{};
 };
